requirements.txt
emoji
geopy
hydra-core
kaleido
loguru
nltk
numpy
pandas
scikit-learn
stanza
plotly
pre-commit
torch
transformers
twikit
vaderSentiment
